Added logging messages. Readded mpic++ library as some compiles require it (the cluster for one)
git-svn-id: https://svn.bcgsc.ca/svn/personal/jsimpson/partitionAssembly/PathWalker/src@38459 ad02020f-2a26-4133-947e-1829317c1416

+++ Parallel/Messages.cpp
@@ -91,7 +91,7 @@ void SeqOpMessage::handle(int senderID, NetworkSequenceCollection& handler)
 //
 void SeqOpMessage::print() const
 {
-	printf("Message type: %d Sequence: %s Operation: %d size: %d\n", (int)Message::m_type, m_seq.decode().c_str(), (int)m_operation, getNetworkSize());
+	printf("Message type: %d Sequence: %s Operation: %d size: %d\n", (int)Message::m_type, m_seq.decode().c_str(), (int)m_operation, (int)getNetworkSize());
 }
